added visible satellites packet, fixed printing of software version to be more friendly.

+++ c/sirf_driver/src/UnknownPacket.cpp
@@ -11,15 +11,15 @@ namespace SiRF {
   void UnknownPacket::outputBinary(std::ostream &out) const {
     // reconstruct the SiRF stream (kind of) we won't bother with
     // the checksum
-    out << 0xa0 << 0xa2 << (uint8)((payload.size() + 1) >> 8)
+    out << (uint8)0xa0 << (uint8)0xa2 << (uint8)((payload.size() + 1) >> 8)
 	<< (uint8)((payload.size() + 1) & 0xff);
     out << type;
     for (std::vector<uint8>::const_iterator itr = payload.begin();
 	 itr != payload.end(); ++itr) {
       out << *itr;
     }
-    out << 0xff << 0xff // no need for a real checksum
-	<< 0xb0 << 0xb3;
+    out << (uint8)0xff << (uint8)0xff // no need for a real checksum
+	<< (uint8)0xb0 << (uint8)0xb3;
   }
   
   // input from a stream
